+++ CHANGELOG.md
@@ -2461,3 +2461,8 @@ Alternately, the _Daily Diary of my Programming Dreams_.
   1.  Added [fiddles/vue/fiddle-0007-PrintPreviewJs](fiddles/vue/fiddle-0007-PrintPreviewJs)
 
 
+* Added [fiddles/vue/fiddle-0008-PlotlyHw](fiddles/vue/fiddle-0008-PlotlyHw)
+* Added [fiddles/vue/fiddle-0008-PlotlyHw](fiddles/vue/fiddle-0008-PlotlyHw)
+* Added [fiddles/vue/fiddle-0008-PlotlyHw](fiddles/vue/fiddle-0008-PlotlyHw)
+* Added [fiddles/vue/fiddle-0008-PlotlyHw](fiddles/vue/fiddle-0008-PlotlyHw)
+* Added [fiddles/vue/fiddle-0008-PlotlyHwJs](fiddles/vue/fiddle-0008-PlotlyHwJs)